Add custom PDF theme
This PDF theme file creates a custom PDF theme, that is as close to the
current ownCloud PDF theme as (appears to be) possible, within the 
constraints of asciidoctor-pdf. While not an exact match, it’s 
pretty close. Personally, I’m not for replicating an existing theme,
when the opportunity may present itself to revise it, and make updates 
and improvements.