Add a PDF theme file
asciidoctor-pdf uses a default theme, which, while fine, isn't enough
for ownCloud's needs. Specifically, it looks quite different to the
PDF's currently being generated. Given that, this commit provides a
custom theme, which overrides the relevant settings, and copies in a lot
of the other settings for two reasons. Firstly, so that the generated
PDF will look as close as practically possible to the existing PDFs.
Secondly, so that we can all know what the supported options are, what a
configuration file looks like, and have an easier way of making changes
in the future, as and when required.