Convert rST source formatting to AsciiDoc
During the conversion from rST to AsciiDoc by Pandoc, the source code
formatting didn't quite change to native AsciiDoc. This change finishes
the migration off, for a section of source code blocks. It's being done
as some aspect of it was corrupting the generation of the table of
contents in the PDF version of the developer manual.